Move interpolate to utils
