+++ training_stats/gpxfile.py
@@ -2,7 +2,6 @@ import re
 import xml.etree.cElementTree as ET
 from datetime import datetime
 
-import numpy as np
 
 
 def get_time(trkpt, namespace):
@@ -43,9 +42,3 @@ def get_hr_measurements(gpx_file):
     data = [(get_time(p, ns), get_hr(p, ns)) for p in trkpts]
     start_time = decode_iso_time(data[0][0])
     return [(convert_time(t, start_time), hr) for (t, hr) in data if hr]
-
-
-def interpolate(points):
-    time, values = zip(*points)
-    new_time = np.arange(int(time[0]), int(time[-1]) + 1)
-    return list(zip(new_time, np.interp(new_time, time, values)))
