+++ training_stats/utils.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+
+def interpolate(points):
+    time, values = zip(*points)
+    new_time = np.arange(int(time[0]), int(time[-1]) + 1)
+    return list(zip(new_time, np.interp(new_time, time, values)))
